feature: set up dictionary models
